集成测试成功

+++ src/main/java/cn/org/hentai/dns/dns/entity/ResourceRecord.java
@@ -15,11 +15,11 @@ public class ResourceRecord
     // public int class;            // 固定值就不声明了
     public int ttl;
     public int dlen;
-    public int ipv4;                // 如果type == TYPE_A，则使用IP作为data，省得转来转去
+    public long ipv4;               // 如果type == TYPE_A，则使用IP作为data，省得转来转去
     public Inet6Address ipv6;       // 同上
     public byte[] data;
 
-    public ResourceRecord(String name, int type, int ttl, int ipv4)
+    public ResourceRecord(String name, int type, int ttl, long ipv4)
     {
         this.name = name;
         this.type = type;
@@ -50,7 +50,7 @@ public class ResourceRecord
     {
         if (type == Message.TYPE_A)
         {
-            int a = (ipv4 >> 24) & 0xff,
+            long a = (ipv4 >> 24) & 0xff,
                     b = (ipv4 >> 16) & 0xff,
                     c = (ipv4 >> 8) & 0xff,
                     d = ipv4 & 0xff;
